pyproject.toml
[project]
name = "mlb-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.3",
    "aiohttp>=3.12.15",
    "httpx>=0.28.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.4",
    "mcp>=1.12.4",
    "python-dotenv>=1.1.1",
]
